fix: Use an AbortSignal for per-request timeouts (#96)
A potential fix for
https://github.com/elastic/elastic-transport-js/issues/63, largely
inspired by a community member's PR that was never merged:
https://github.com/elastic/elastic-transport-js/pull/55

According to an Undici core committer in this comment
https://github.com/elastic/elasticsearch-js/issues/1716#issuecomment-1167173492
the issue that triggers the MaxListenersExceededWarning, and possibly a
memory leak in some cases, is caused by attaching an EventEmitter to
each request by default when a per-request timeout is set, rather than
attaching an AbortSignal.

My assumption is that an EventEmitter was used because AbortSignal and
AbortController were not added to Node.js until v14.17.0, so we couldn't
guarantee v14 users would have it. I'm not certain why using
EventEmitters makes a difference memory-wise, but it does get rid of the
MaxListenersExceededWarning.